Prevent adding sensitive cluster certs to state
This is supposed to improve our security posture some more, also knowing
that once the certificates in question, which allow admin access to the
cluster, have leaked, there's no way to rotate them and we need to
recreate such the affected cluster (as happened in the past). Now should
the state file leak one way or the other, at least we're not giving away
admin access to the cluster anymore.

=>
https://github.com/opentelekomcloud/terraform-provider-opentelekomcloud/issues/2208